moved SelectedItems into tools > move
