+++ src/main/java/com/marginallyclever/robotoverlord/tools/move/RotateEntityToolOneAxis.java
@@ -8,7 +8,6 @@ import com.marginallyclever.robotoverlord.Entity;
 import com.marginallyclever.robotoverlord.Viewport;
 import com.marginallyclever.robotoverlord.components.PoseComponent;
 import com.marginallyclever.robotoverlord.tools.EditorTool;
-import com.marginallyclever.robotoverlord.tools.SelectedItems;
 
 import javax.vecmath.Matrix4d;
 import javax.vecmath.Point3d;
@@ -16,6 +15,7 @@ import javax.vecmath.Tuple3d;
 import javax.vecmath.Vector3d;
 import java.awt.event.KeyEvent;
 import java.awt.event.MouseEvent;
+import java.util.List;
 
 public class RotateEntityToolOneAxis implements EditorTool {
     /**
@@ -79,10 +79,7 @@ public class RotateEntityToolOneAxis implements EditorTool {
      * which axis of rotation will be used?  0,1,2 = x,y,z
      */
     private int rotation=2;
-
     private boolean hovering = false;
-
-    private double angleAtStartOfDrag;
 
     /**
      * This method is called when the tool is activated. It receives the SelectedItems object containing the selected
@@ -91,8 +88,8 @@ public class RotateEntityToolOneAxis implements EditorTool {
      * @param selectedItems The selected items to be manipulated by the tool.
      */
     @Override
-    public void activate(SelectedItems selectedItems) {
-        this.selectedItems = selectedItems;
+    public void activate(List<Entity> list) {
+        this.selectedItems = new SelectedItems(list);
         if (selectedItems.isEmpty()) return;
 
         setPivotMatrix(EditorUtils.getLastItemSelectedMatrix(selectedItems));
@@ -140,7 +137,6 @@ public class RotateEntityToolOneAxis implements EditorTool {
             dragging = true;
             hovering = true;
             startPoint = EditorUtils.getPointOnPlaneFromCursor(MatrixHelper.getXYPlane(startMatrix),viewport,event.getX(), event.getY());
-            angleAtStartOfDrag = getAngleBetweenPoints(startPoint);
             if(selectedItems!=null) selectedItems.savePose();
         }
     }
