+++ src/main/java/com/marginallyclever/robotoverlord/tools/move/TranslateEntityToolTwoAxis.java
@@ -8,13 +8,13 @@ import com.marginallyclever.robotoverlord.Entity;
 import com.marginallyclever.robotoverlord.Viewport;
 import com.marginallyclever.robotoverlord.components.PoseComponent;
 import com.marginallyclever.robotoverlord.tools.EditorTool;
-import com.marginallyclever.robotoverlord.tools.SelectedItems;
 
 import javax.vecmath.Matrix4d;
 import javax.vecmath.Point3d;
 import javax.vecmath.Vector3d;
 import java.awt.event.KeyEvent;
 import java.awt.event.MouseEvent;
+import java.util.List;
 
 public class TranslateEntityToolTwoAxis implements EditorTool {
     private final double padSize = 1;
@@ -54,8 +54,8 @@ public class TranslateEntityToolTwoAxis implements EditorTool {
     private boolean hovering = false;
 
     @Override
-    public void activate(SelectedItems selectedItems) {
-        this.selectedItems = selectedItems;
+    public void activate(List<Entity> list) {
+        this.selectedItems = new SelectedItems(list);
         if (selectedItems.isEmpty()) return;
 
         setPivotMatrix(EditorUtils.getLastItemSelectedMatrix(selectedItems));
